pyproject.toml
[tool.poetry]
name="best-stock-analyst" 
version="0.1.0"
description='Simple stock event planner'
authors=["Venkatesh I"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.12"
unstructured = "==0.10.25"
pyowm = "3.3.0"
python-dotenv = "1.0.0"
ipykernel = "^6.29.5"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
pydantic = ">=1,<3"
phidata = "^2.7.9"
openai = "^1.60.0"
groq = "^0.15.0"
duckduckgo-search = "^7.2.1"
yfinance = "^0.2.52"
streamlit = "^1.41.1"


[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']# No

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"